Mark current high-level ACP API as V1
This will allow us to work on changes to make it compatible with
upcoming breaking changes to the ACP proposal in parallel without
breaking existing apps before the changes are live.
